requirements.txt
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
pathlib2>=2.3.7
urllib3>=2.0.0
lxml>=4.9.0 